Modify the code for solving the feedback gain matrix and change the input variables to double type to improve accuracy.

+++ program/math/riccati_solver.h
@@ -31,7 +31,7 @@ void project(basic_status_t current, basic_status_t *project, float v, float t, 
 void lqr_control_test(unsigned short index, int angle_ctrl, float angle_yaw, float distance_n, float distance_e);
 unsigned char lqr_control(unsigned short index, basic_status_t status);
 
-void solve_feedback_value(float p[3][3], float a[3][3], float b[3][2], float x[3][1], float r, float fb[2][1]);
-void solve_riccati_equation(float a[3][3], float b[3][2], float q, float r, float p[3][3]);
+void solve_feedback_value(float p[3][3], double a[3][3], double b[3][2], double x[3][1], float r, float fb[2][1]);
+void solve_riccati_equation(double a[3][3], double b[3][2], float q, float r, float p[3][3]);
 
 #endif
